ataque
dealDamage
