--- Assets/Scenes/PruebaUnidades.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d15bbad5760497849b28ff92c23c09f5
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
