+++ Assets/ScriptableObjects/Units/EnemyLargeCrocodileUnit.asset
@@ -18,9 +18,9 @@ MonoBehaviour:
   unitAttackType: 0
   cost: 0
   reward: 500
-  initialLife: 1000
-  currentLife: 1000
+  life: 1000
   attack: 50
   rangeAttack: 1
   rateFire: 2
   speedMovement: 1
+  description:
