+++ Assets/ScriptableObjects/Units/EnemyMediumUnit.asset
@@ -18,9 +18,9 @@ MonoBehaviour:
   unitAttackType: 0
   cost: 0
   reward: 100
-  initialLife: 150
-  currentLife: 150
+  life: 150
   attack: 10
   rangeAttack: 1
   rateFire: 1
   speedMovement: 1
+  description:
