+++ Assets/ScriptableObjects/Units/TestBasicAllyUnit.asset
@@ -14,11 +14,13 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   unitSize: 0
   unitFaction: 0
+  unitSpecialSkill: 0
+  unitAttackType: 0
   cost: 50
   reward: 0
-  initialLife: 50
-  currentLife: 50
+  life: 50
   attack: 10
   rangeAttack: 2
   rateFire: 3
   speedMovement: 3
+  description:
