+++ Assets/ScriptableObjects/Units/TestBasicEnemyUnit.asset
@@ -14,11 +14,13 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   unitSize: 0
   unitFaction: 1
+  unitSpecialSkill: 0
+  unitAttackType: 0
   cost: 0
   reward: 50
-  initialLife: 50
-  currentLife: 50
+  life: 50
   attack: 10
   rangeAttack: 1
   rateFire: 1
   speedMovement: 3
+  description:
