--- Assets/Scripts/UnitBehavior/Antiguo TestAlly.txt.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 98399a7c3bd59e44cbb1284b4761fd81
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
