+++ Assets/Scripts/UnitBehavior/UnitStats.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: bc4a1e20e2cae504e96b3ba751f30edf
+guid: a3508f0e3e758a048b68b6258af68e21
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
