+++ ProjectSettings/TagManager.asset
@@ -7,6 +7,7 @@ TagManager:
   - Ally
   - Enemy
   - EnemyBase
+  - AllyBase
   layers:
   - Default
   - TransparentFX
